import of CINT 5.14.87. For what is new see http://root.cern.ch/root/Cint.phtml?relnote.
git-svn-id: http://root.cern.ch/svn/root/trunk@2189 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/include/mkincld.c
@@ -316,7 +316,7 @@ int gen_time()
 #endif
 
 #ifdef CLK_TCK
-  fprintf(fp,"#define \tCLK_TCK (%d)\n",CLK_TCK);
+  fprintf(fp,"#define \tCLK_TCK (%ld)\n",(long)CLK_TCK);
 #endif
 
   fprintf(fp,"#endif\n");
